Execution_Engine/requirements.txt
requests
redis
fastapi
uvicorn
docker

# Core Data Handling
numpy
pandas

# Technical Analysis (for OHLCV calculations)
ta
pandas-ta

# Data Fetching (if needed)
yfinance
ccxt
requests

# Utilities
python-dotenv
pytz
python-dateutil
groq
psutil